수정 기능 보완

+++ src/main/java/com/example/ch8_5/to/BoardTo.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class BoardTo {
-    private int bno;
+    private Long bno;
     private String title;
     private String writer;
     private String content;
